+++ Assets/Prefabs/Pizza.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 1928732646584446961}
   - component: {fileID: 4860670117560215006}
   - component: {fileID: 1115104595579471920}
-  - component: {fileID: 7373525149685284885}
   m_Layer: 0
   m_Name: Pizza
   m_TagString: Untagged
@@ -145,15 +144,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   selectedObject: {fileID: 0}
---- !u!114 &7373525149685284885
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5187429636330366291}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ec9db77e03b5f1d459e5ead9885185d9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
