--- Assets/Scripts/PizzaResetBehaviour.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec9db77e03b5f1d459e5ead9885185d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
